Exception Handling and Paginating completed

+++ src/main/java/kz/halykacademy/bookstore/service/authors/AuthorServiceImpl.java
@@ -2,6 +2,7 @@ package kz.halykacademy.bookstore.service.authors;
 
 import kz.halykacademy.bookstore.dao.authors.AuthorEntity;
 import kz.halykacademy.bookstore.dao.authors.AuthorRepository;
+import kz.halykacademy.bookstore.web.ExceptionHandling.ResourceNotFoundException;
 import kz.halykacademy.bookstore.web.authors.Author;
 import kz.halykacademy.bookstore.web.authors.SaveAuthor;
 import org.hibernate.tool.schema.internal.exec.ScriptTargetOutputToFile;
@@ -9,28 +10,19 @@ import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 import java.util.List;
+import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
 
-
-
-
 /**
- *
- *
- *
- *              Допиши Везьде Exception Handling!!!
- *
- *              Внедрить Pageble!
- *
- *              Внедрить Exception Handling!
- *
- *              Поменять openApi спецификацию
- *
- *
- *
- * */
-
+ * Допиши Везьде Exception Handling!!!
+ * <p>
+ * Внедрить Pageble!
+ * <p>
+ * Внедрить Exception Handling!
+ * <p>
+ * Поменять openApi спецификацию
+ */
 
 
 @Service
@@ -53,18 +45,16 @@ public class AuthorServiceImpl implements AuthorService {
 
     @Override
     public Author getAuthor(long id) {
-//        return authorRepository.findById(id)
-//                .map(AuthorEntity::toDto)
-//                .orElseThrow((Supplier<Throwable>) () -> {
-//                    new ResourceNotFoundException();
-//                });
-
-        return authorRepository.getReferenceById(id).toDto();
+        return authorRepository.findById(id)
+                .map(AuthorEntity::toDto)
+                .orElseThrow(() -> new ResourceNotFoundException("Author not found. Invalid id supplied!"));
     }
 
     @Transactional
     @Override
     public Author putAuthor(long id, SaveAuthor author) {
+        if (!authorRepository.existsById(id))
+            throw new ResourceNotFoundException("Author not found. Invalid id supplied!");
         authorRepository.updateAuthorById(
                 id,
                 author.getFirstName(),
@@ -93,6 +83,8 @@ public class AuthorServiceImpl implements AuthorService {
 
     @Override
     public void deleteAuthor(long id) {
+        if (!authorRepository.existsById(id))
+            throw new ResourceNotFoundException("Author not found. Invalid id supplied!");
         authorRepository.deleteById(id);
     }
